+++ src/core/integral_fink.hpp
@@ -76,7 +76,7 @@ struct FinkFCIDUMP : FCIDUMP<double> {
             fd->ts[s].clear();
             for (uint16_t i = 0; i < n; i++)
                 for (uint16_t j = 0; j < n; j++)
-                    fd->ts[s](i, j) = t(s, i, j);
+                    fd->ts[s](i, j) = t((uint8_t)s, i, j);
         }
         for (size_t s = 0; s < vgs.size(); s++) {
             for (uint16_t i = 0; i < n; i++)
